Practicum/P2/YkeaWebsite/requirements.txt
Django==1.11.12
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
django-heroku==0.3.1
Pillow==5.1.0
gunicorn==19.8.1
whitenoise==3.3.1
psycopg2==2.7.4
